+++ Tests/Runtime/SimpleTests.cs
@@ -1,6 +1,5 @@
 using System;
 using NUnit.Framework;
-using SaveSystem.Internal;
 using SaveSystem.Tests.TestObjects;
 using UnityEngine;